Decouple isValidKioskMode from KioskContext's React context module
isValidKioskMode was defined inside common/contexts/KioskContext/index.tsx,
alongside the KioskContext/KioskProvider React machinery. AppContext and
civic-uk.ts both import it via the KioskContext barrel for a generic
cookie-consent check, unrelated to kiosk mode itself - but importing from a
barrel pulls in the whole module, so anything else added to
KioskContext/index.tsx (e.g. HistoryProvider, in a following commit) rides
along into every app that transitively imports AppContext or civic-uk.ts,
including identity, which never uses kiosk mode at all.

Move isValidKioskMode into its own tiny module with no dependency on React or
the rest of KioskContext, and have AppContext/civic-uk.ts import it directly.
