+++ common/contexts/KioskContext/index.tsx
@@ -18,25 +18,6 @@ import {
   ReadingRoomStories,
   TendernessAndRageContent,
 } from '@weco/common/server-data/prismic';
-import { KioskModeOptionId } from '@weco/toggles';
-import toggleConfig from '@weco/toggles/toggles';
-
-// Valid kiosk mode IDs extracted from toggles config
-const VALID_KIOSK_MODE_IDS =
-  toggleConfig.modes
-    .find(mode => mode.id === 'kioskMode')
-    ?.options.map(option => option.id) ?? [];
-
-/**
- * Validates that a cookie value is a valid kiosk mode option ID.
- * Use this server-side to validate the toggle_kioskMode cookie before using it.
- */
-export function isValidKioskMode(value: unknown): value is KioskModeOptionId {
-  return (
-    typeof value === 'string' &&
-    (VALID_KIOSK_MODE_IDS as readonly string[]).includes(value)
-  );
-}
 
 export type KioskContextType = {
   isKiosk: boolean;
